+++ Cod3rsGrowth.sln
@@ -18,8 +18,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cod3rsGrowth.Dominio", "Cod
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cod3rsGrowth.Servicos", "Cod3rsGrowth.Servicos\Cod3rsGrowth.Servicos.csproj", "{DDED1034-8923-4F06-ABE6-027BCA898CF7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cod3rsGrowth.API", "Cod3rsGrowth.API\Cod3rsGrowth.API.csproj", "{20C862CF-09A7-43F3-8EAE-51E10103D269}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -50,10 +48,6 @@ Global
 		{DDED1034-8923-4F06-ABE6-027BCA898CF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DDED1034-8923-4F06-ABE6-027BCA898CF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DDED1034-8923-4F06-ABE6-027BCA898CF7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{20C862CF-09A7-43F3-8EAE-51E10103D269}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{20C862CF-09A7-43F3-8EAE-51E10103D269}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{20C862CF-09A7-43F3-8EAE-51E10103D269}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{20C862CF-09A7-43F3-8EAE-51E10103D269}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
